requirements.txt
uvicorn
python-dotenv
requests
openai-agents
FastMCP